athletics_event_platform_backend/requirements.txt
Django==6.0.1
djangorestframework==3.16.1
django-cors-headers==4.9.0
gunicorn==25.1.0
psycopg2-binary==2.9.11
dj-database-url==3.1.2
whitenoise==6.11.0
python-dotenv==1.0.0
